# Conflicts:
#	files/mcloud/appium-base-driver/build/lib/basedriver/helpers.js

+++ files/mcloud/appium-base-driver/build/lib/basedriver/helpers.js
@@ -385,8 +385,11 @@ async function configureApp(app, supportedAppExtensions) {
         } finally {
           if (localAppsFolder != undefined) {
             _logger.default.info(`[MCLOUD] Going to remove lock file ${lockFile}`);
-
-            await _appiumSupport.fs.unlink(lockFile);
+            if (await _appiumSupport.fs.exists(lockFile)) {
+              await _appiumSupport.fs.unlink(lockFile);
+            } else {
+              _logger.default.warn(`[MCLOUD] Lock file ${lockFile} was not found. Probably it was removed by another thread which was downloading app in parallel`);
+            }
           }
         }
       }
